refactor: Export date-related components from default indices since tree-shaking is now working properly
