+++ packages/components/src/Form/Fields/index.ts
@@ -28,6 +28,8 @@ export * from './FieldCheckbox'
 export * from './FieldCheckboxGroup'
 export * from './FieldChips'
 export * from './FieldColor'
+export * from './FieldDate'
+export * from './FieldDateRange'
 export * from './FieldRadio'
 export * from './FieldRadioGroup'
 export * from './FieldRangeSlider'
