+++ packages/components/src/Form/Inputs/index.ts
@@ -31,6 +31,8 @@ export * from './InlineInputText'
 export * from './InlineTextArea'
 export * from './InputChips'
 export * from './InputColor'
+export * from './InputDate'
+export * from './InputDateRange'
 export * from './InputFilters'
 export * from './InputHidden'
 export * from './InputSearch'
