+++ www/src/documentation/components/content/calendar.mdx
@@ -2,20 +2,8 @@
 title: Calendar
 ---
 
-import { MessageBar, Code } from '@looker/components'
+import { Code } from '@looker/components'
 import { CalendarPropTable } from './demos'
-
-<MessageBar my="large" noActions>
-  Note: Calendar must be imported from its specific location rather than the
-  package root.
-</MessageBar>
-
-Import path: `@looker/components/lib/Calendar`
-
-<MessageBar noActions>
-  If you need an interactive date picker, use{' '}
-  <a href="../forms/input-date">InputDate</a> or InputDateRange instead.
-</MessageBar>
 
 `Calendar` is a stateless component which renders a standard calendar month. It accepts a limited amount of props, and is the foundation of our more useful [`InputDate`](/components/forms/input-date) and `InputDateRange` components.
