+++ www/src/documentation/components/content/date-format.mdx
@@ -3,15 +3,7 @@ title: DateFormat
 github: DateTimeFormat/DateFormat.tsx
 ---
 
-import { MessageBar } from '@looker/components'
 import { DateFormatTable, LocaleTable } from './demos'
-
-<MessageBar my="large" noActions>
-  Note: DateFormat must be imported from its specific location rather than the
-  package root.
-</MessageBar>
-
-Import path: `@looker/components/lib/DateFormat`
 
 As it's name implies, `DateFormat` is used to format a javascript Date object to human readable and localized text.
