+++ www/src/documentation/components/content/time-format.mdx
@@ -3,15 +3,7 @@ title: TimeFormat
 github: DateTimeFormat/TimeFormat.tsx
 ---
 
-import { MessageBar } from '@looker/components'
 import { TimeFormatTable, LocaleTable } from './demos'
-
-<MessageBar my="large" noActions>
-  Note: TimeFormat must be imported from its specific location rather than the
-  package root.
-</MessageBar>
-
-Import path: `@looker/components/lib/TimeFormat`
 
 As it's name implies, `TimeFormat` is used to format a javascript Date object to human readable and localized text.
