+++ www/src/documentation/components/forms/input-date-range.mdx
@@ -4,15 +4,6 @@ github: Form/Inputs/InputDateRange/InputDateRange.tsx
 storybook: true
 ---
 
-import { MessageBar } from '@looker/components'
-
-<MessageBar my="large" noActions>
-  Note: InputDatRange must be imported from its specific location rather than
-  the package root.
-</MessageBar>
-
-Import path: `@looker/components/lib/InputDateRange`
-
 `InputDateRange` provides an interface to select to a date range, either through the calendar interface or text input boxes.
 
 ## Value
