+++ www/src/documentation/components/forms/input-date.mdx
@@ -4,15 +4,6 @@ github: Form/Inputs/InputDate/InputDate.tsx
 storybook: true
 ---
 
-import { MessageBar } from '@looker/components'
-
-<MessageBar my="large" noActions>
-  Note: InputDate must be imported from its specific location rather than the
-  package root.
-</MessageBar>
-
-Import path: `@looker/components/lib/InputDate`
-
 `InputDate` provides a text input and calendar to select a single date, which can be used as a form input or filter.
 
 ## Value & Default Value
